refactor(small main): move back into internal
Go best practices state the main.go directory should stay small
