+++ internal/kubernetes/finder_test.go
@@ -1,4 +1,4 @@
-package main
+package kubernetes
 
 import (
 	// standard packages
