+++ internal/kubernetes/watcher_test.go
@@ -1,4 +1,4 @@
-package main
+package kubernetes
 
 import (
 	// standard packages
@@ -10,7 +10,6 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	// internal packages
-	kubeInternal "volume-cleaner/internal/kubernetes"
 	structInternal "volume-cleaner/internal/structure"
 	testInternal "volume-cleaner/internal/tests"
 )
@@ -48,7 +47,7 @@ func TestWatcherLabelling(t *testing.T) {
 		time.Sleep(2 * time.Second)
 
 		// no pvc should have labels right now
-		pvcs := kubeInternal.PvcList(kube, "test")
+		pvcs := PvcList(kube, "test")
 
 		_, ok := pvcs[0].Labels["volume-cleaner/unattached-time"]
 		assert.Equal(t, ok, false)
@@ -65,7 +64,7 @@ func TestWatcherLabelling(t *testing.T) {
 
 		// should be no change
 
-		pvcs = kubeInternal.PvcList(kube, "test")
+		pvcs = PvcList(kube, "test")
 
 		_, ok = pvcs[0].Labels["volume-cleaner/unattached-time"]
 		assert.Equal(t, ok, false)
@@ -82,7 +81,7 @@ func TestWatcherLabelling(t *testing.T) {
 
 		// should have new labels
 
-		pvcs = kubeInternal.PvcList(kube, "test")
+		pvcs = PvcList(kube, "test")
 
 		_, ok = pvcs[0].Labels["volume-cleaner/unattached-time"]
 		assert.Equal(t, ok, true)
@@ -116,7 +115,7 @@ func TestInitialScan(t *testing.T) {
 		}
 
 		// no pvc should have labels right now
-		pvcs := kubeInternal.PvcList(kube, "test")
+		pvcs := PvcList(kube, "test")
 
 		_, ok := pvcs[0].Labels["volume-cleaner/unattached-time"]
 		assert.Equal(t, ok, false)
@@ -138,7 +137,7 @@ func TestInitialScan(t *testing.T) {
 
 		// should have new labels
 
-		pvcs = kubeInternal.PvcList(kube, "test")
+		pvcs = PvcList(kube, "test")
 
 		_, ok = pvcs[0].Labels["volume-cleaner/unattached-time"]
 		assert.Equal(t, ok, true)
